requirements.txt
Django
djangorestframework
loguru
pillow
drf-yasg